CI: build and push each architecture with a different version label
Use a matrix to append -$arch to the version when uploading, while
keeping original multi-arch image intact. Unfortunately, we need
multiple jobs to do this, because we need different values for the
"tags" argument of docker/build-push-action.

We need this because there's no way to specify the architecture when
using a github action in a step.